+++ packages/govuk-frontend/src/govuk/components/file-upload/file-upload.mjs
@@ -238,7 +238,7 @@ export class FileUpload extends ConfigurableComponent {
     // so we first need to make sure it's a `Node`
     if (event.target instanceof Node) {
       if (this.$root.contains(event.target)) {
-        if (event.dataTransfer && isContainingFiles(event.dataTransfer)) {
+        if (event.dataTransfer && this.canDrop(event.dataTransfer)) {
           // Only update the class and make the announcement if not already visible
           // to avoid repeated announcements on NVDA (2024.4) + Firefox (133)
           if (
@@ -291,7 +291,7 @@ export class FileUpload extends ConfigurableComponent {
   onDrop(event) {
     event.preventDefault()
 
-    if (event.dataTransfer && isContainingFiles(event.dataTransfer)) {
+    if (event.dataTransfer && this.canFillInput(event.dataTransfer)) {
       this.$input.files = event.dataTransfer.files
 
       // Dispatch a `change` event so external code that would rely on the `<input>`
@@ -301,6 +301,58 @@ export class FileUpload extends ConfigurableComponent {
 
       this.hideDraggingState()
     }
+  }
+
+  /**
+   * Confirms if enhanced `<input>` can be filled with files from the given `DataTransfer`
+   *
+   * @param {DataTransfer} dataTransfer - The `DataTransfer` being dropped
+   * @returns {boolean} - true if the `DataTransfer` contains files, in number matching the `multiple` attribute of the original `<input>`
+   * @private
+   */
+  canFillInput(dataTransfer) {
+    return this.matchesInputCapacity(dataTransfer.files.length)
+  }
+
+  /**
+   * Confirms if the content of a `DataTransfer` dragged over component can be dropped
+   *
+   * Unfortunately, there's a certain level of uncertainty in Safari which does not
+   * even provide a list of `items` while dragging (and seems to even miss the `types` sometimes)
+   *
+   * @param {DataTransfer} dataTransfer - The `DataTransfer` being dragged
+   * @returns {boolean} - true if the `DataTransfer` looks OK for filling the input, false otherwise
+   * @private
+   */
+  canDrop(dataTransfer) {
+    // If the browser is kind enough to give a list of items, we'll use that as source of truth
+    if (dataTransfer.items.length) {
+      return this.matchesInputCapacity(countFileItems(dataTransfer.items))
+    }
+
+    // If we have some type information, we'll use that
+    if (dataTransfer.types.length) {
+      return dataTransfer.types.includes('Files')
+    }
+
+    // If we have nothing to go by, we'll assume things are OK
+    // until we have a more accurate picture inside the `drop` event
+    return true
+  }
+
+  /**
+   * Confirms the given number of files matches that allowed by the enhanced `<input>`
+   *
+   * @param {number} numberOfFiles - The number of files
+   * @returns {boolean} - `true` if the enhanced `<input>` can accept that number of files
+   * @private
+   */
+  matchesInputCapacity(numberOfFiles) {
+    if (this.$input.multiple) {
+      return numberOfFiles > 0
+    }
+
+    return numberOfFiles === 1
   }
 
   /**
@@ -440,22 +492,22 @@ export class FileUpload extends ConfigurableComponent {
 }
 
 /**
- * Checks if the given `DataTransfer` contains files
+ * Counts the number of `DataTransferItem` whose kind is `file`
  *
- * @internal
- * @param {DataTransfer} dataTransfer - The `DataTransfer` to check
- * @returns {boolean} - `true` if it contains files or we can't infer it, `false` otherwise
+ * @param {DataTransferItemList} list - The list
+ * @returns {number} - The number of items whose kind is `file` in the list
  */
-function isContainingFiles(dataTransfer) {
-  // Safari sometimes does not provide info about types :'(
-  // In which case best not to assume anything and try to set the files
-  const hasNoTypesInfo = dataTransfer.types.length === 0
+function countFileItems(list) {
+  let result = 0
 
-  // When dragging images, there's a mix of mime types + Files
-  // which we can't assign to the native input
-  const isDraggingFiles = dataTransfer.types.some((type) => type === 'Files')
-
-  return hasNoTypesInfo || isDraggingFiles
+  // `DataTransferItemList` is not iterable
+  // eslint-disable-next-line @typescript-eslint/prefer-for-of
+  for (let i = 0; i < list.length; i++) {
+    if (list[i].kind === 'file') {
+      result++
+    }
+  }
+  return result
 }
 
 /**